Add files via upload
The method we have used in this project is called Behavior Cloning using NVIDIA model.
Behavior cloning in machine learning is a technique where a model learns to imitate the actions
of an expert by observing examples of those actions. From this method we have used Deep
neural network, for feature extraction we have used Convolutional Neural Network (CNN) and
for the last part, usually behavior cloning is a classification problem but for our case it is a
continuous regression.